requirements.txt
watchdog==6.0.0
cloudinary==1.44.1
python-dotenv==1.2.1
twilio==9.10.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.4
frozenlist==1.8.0
idna==3.11
multidict==6.7.1
propcache==0.4.1
PyJWT==2.11.0
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.22.0